Correct the pfentry code
    Re-reading the ASM, the pfentry.S file provides a lot of the
behavior which we were attempting to provide in trap.c. For instance, it
takes care of dispatching to the user trap handler, and putting the
required pointer on the stack.

    Conseqently patched trap.c to only provide required behavior,
patched set_trap_handler routine to correctly set the pfentry trap
handler as the user trap handler instead of setting the user trap
handler directly.

    Also patched pfentry.S to correct a stack indexing bug where we
attempted to fetch data at negative indicies from the stack pointer
instead of the corect and intended positive indicies. Freaking
down-growing stack....