Add preemption-preventing syscall
  Add a system call which prevents a process from being preempted N
times where N is an unsigned integer. The idea here is that it allows me
to improve the quality of my print-statement based logging by making my
DEBUGs atomic with respect to the scheduler.

  At present this routine constitutes a security hole because it does
not prevent a process from setting its preemption escape while escaped.
This allows a process (such as a system task) to hard wire itself to a
CPU by preventing the scheduler from kicking it from that CPU.